// src/api/admin/subcategory/type.ts
export type TSubcategory = {
  name: string;
  categoryId: string;
};

export type TSubcategoryPut = {
  subcategoryId: string;
} & TSubcategory;

export type TSubcategoryRes = {
  _id: string;
  name: string;
  categoryId: {
    _id: string;
    name: string;
  };
  createdAt: Date;
  deletedAt: Date | null;
  updatedAt: Date | null;
  createdBy: string | null; // Referencing User
  deletedBy: string | null; // Referencing User or null
  updatedBy: string | null; // Referencing User or null
};